Factor out session validity formatting
